An S4 class for an antsImageIterator

C++ type used to represent an ITK image iterator. the actual iterator is of C++ type 'itk::ImageRegionIteratorWithIndex< itk::image< pixeltype, dimension > >'

# S4 method for antsImageIterator
show(object)

# S4 method for antsImageIterator
initialize(.Object, pixeltype = "float",
  dimension = 3, components = 1)

Arguments

object
input object to convert

.Object
input object to convert

pixeltype
string e.g. "float" "unsigned char" "int"

dimension
dimensionality of the image

components
number of components per pixel

Slots

pixeltype

usually float, can be other types unsigned char, int, double etc noting that short is not supported

dimension

usually 2 or 3 but can be 4

components

number of pixel components, currently only 1 is suppored

pointer

to the memory location of the itk object