Denoise an image

Denoise an image using a spatially adaptive filter originally described in J. V. Manjon, P. Coupe, Luis Marti-Bonmati, D. L. Collins, and M. Robles. Adaptive Non-Local Means Denoising of MR Images With Spatially Varying Noise Levels, Journal of Magnetic Resonance Imaging, 31:192-203, June 2010.

denoiseImage(img, mask, shrinkFactor = 1, p = 1, r = 3,
  noiseModel = c("Rician", "Gaussian"), verbose = FALSE)

Arguments

img
scalar image to denoise.

mask
optional to limit the denoise region.

shrinkFactor
downsampling level performed within the algorithm.

p
patch radius for local sample.

r
search radius from which to choose extra local samples.

noiseModel
either Rician or Gaussian.

verbose
boolean

Value

antsImage denoised version of image

Examples

img <- antsImageRead( getANTsRData("rand")  ) %>% resampleImage( c(32, 32) )
dimg <- denoiseImage( img, img * 0 + 1 )
dimg2 <- denoiseImage( img)