Apply transforms to images.

Apply a transform list to map an image from one domain to another. In image registration, one computes mappings between (usually) pairs of images. These transforms are often a sequence of increasingly complex maps, e.g. from translation, to rigid, to affine to deformation. The list of such transforms is passed to this function to interpolate one image domain into the next image domain, as below. The order matters strongly and the user is advised to familiarize with the standards established in examples.

antsApplyTransforms(fixed, moving, transformlist = "",
  interpolator = c("linear", "nearestNeighbor", "multiLabel", "gaussian",
  "bSpline", "cosineWindowedSinc", "welchWindowedSinc",
  "hammingWindowedSinc", "lanczosWindowedSinc", "genericLabel"),
  imagetype = 0, whichtoinvert = NA, compose = NA, verbose = FALSE,
  ...)

Arguments

fixed
fixed image defining domain into which the moving image is transformed.

moving
moving image to be mapped to fixed space.

transformlist
character vector of transforms generated by antsRegistration where each transform is a filename.

interpolator
Choice of interpolator. Supports partial matching.

linear

nearestNeighbor

multiLabel for label images but genericlabel is preferred

gaussian

bSpline

cosineWindowedSinc

welchWindowedSinc

hammingWindowedSinc

lanczosWindowedSinc

genericLabel use this for label images

imagetype
choose 0/1/2/3 mapping to scalar/vector/tensor/time-series

whichtoinvert
optional list of booleans, same length as transforms. whichtoinvert[i] is TRUE if transformlist[i] is a matrix, and the matrix should be inverted. If transformlist[i] is a warp field, whichtoinvert[i] must be FALSE.

If the transform list is a matrix followed by a warp field, whichtoinvert defaults to c(TRUE,FALSE). Otherwise it defaults to rep(FALSE, length(transformlist)).

compose
if it is a character string pointing to a valid file location, this will force the function to return a composite transformation filename.

verbose
print command and run verbose application of transform.

...
extra parameters

Value

an antsImage or transformation filename is output. 1 -- Failure

See also

Examples

# will give the full form of help
antsApplyTransforms("-h")

#> NULL

# see antsRegistration
# example 1 - simplified
fixed <- antsImageRead( getANTsRData("r16") ,2)
moving <- antsImageRead( getANTsRData("r64") ,2)
fixed <- resampleImage(fixed,c(64,64),1,0)
moving <- resampleImage(moving,c(68,68),1,0)
mytx <- antsRegistration(fixed=fixed , moving=moving ,
  typeofTransform = c("SyN") )

#> stoi: no conversion

#> Error in antsRegistration(fixed = fixed, moving = moving, typeofTransform = c("SyN")): Registration failed. Use verbose mode to diagnose.

mywarpedimage <- antsApplyTransforms( fixed=fixed,moving=moving,
  transformlist=mytx$fwdtransforms )

#> Error in inherits(x, "antsImage"): object 'mytx' not found

testthat::expect_true(antsImagePhysicalSpaceConsistency(mywarpedimage, fixed))

#> Error in check_ants(img1): object 'mywarpedimage' not found

invwarped_image <- antsApplyTransforms( fixed=moving,moving=fixed,
  transformlist=mytx$invtransforms )

#> Error in inherits(x, "antsImage"): object 'mytx' not found

testthat::expect_true(antsImagePhysicalSpaceConsistency(invwarped_image, moving))

#> Error in check_ants(img1): object 'invwarped_image' not found

# full access via listing the inputs in standard ANTs format

res1 <- antsApplyTransforms( fixed=fixed,moving=moving,
  transformlist=mytx$fwdtransforms[2],
  whichtoinvert =1, verbose = TRUE)

#> Error in inherits(x, "antsImage"): object 'mytx' not found

cfile <- antsApplyTransforms( fixed=fixed,moving=moving,
  transformlist=mytx$fwdtransforms,
  compose = tempfile() )

#> Error in inherits(x, "antsImage"): object 'mytx' not found

cimg = antsImageRead(cfile)

#> Error in antsImageRead(cfile): object 'cfile' not found

cout <- antsApplyTransforms( fixed=fixed,moving=moving,
  transformlist=cimg)

#> Error in inherits(x, "antsImage"): object 'cimg' not found

testthat::expect_error(
antsApplyTransforms( fixed=fixed,moving=moving,
  transformlist=cimg, whichtoinvert =1), "nnot invert transform"
)

#> Error: `antsApplyTransforms(...)` threw an error with unexpected message.
#> Expected match: "nnot invert transform"
#> Actual message: "object 'cimg' not found"

testthat::expect_error(
antsApplyTransforms( fixed=fixed,moving=moving,
  transformlist=cimg, whichtoinvert =c(1,2)), "same length"
)

#> Error: `antsApplyTransforms(...)` threw an error with unexpected message.
#> Expected match: "same length"
#> Actual message: "object 'cimg' not found"

testthat::expect_error(antsApplyTransforms( fixed=fixed,moving=moving,
  transformlist= "")
 )